refactor (informationmodel-cat): Move responsibility of publisher lookup to InformationModelFactory
